Move ServerFormDetailsFetcher logic to use case
